Make a lot of changes & enhancements to Dashboard
So now, PaymentHistory component has animated changing of width which is
only animated when user clicks on change fullscreen mode button. When
its metrics are being changed due to resize or smth like that, width and
other properties are not animated. The same is with TransactionHistory
component. At first, it should be server component, but to animate it,
I had to change its metrics and so on. I also enhanced date generating
algorithm, so now positions of date elems look more fancy. And I
enhanced random num generator function, so the diff between nums is more
smooth and resulting graph looks more natural.