Refactor some components and create tests
So, I've fully covered PaymentHistory itself with tests, except
FullScreen icon, because it needs to be mocked to be used in tests, so
it's not possible to determine something from it. The only thing that
is left - e2e and snapshot testing. E2e I will do later, because I don't
know Cypress for now, what about snapshot, I'll do it with the whole
PaymentAndTransactionWrapper after I cover with tests all components
that it consists of. I've partly refactored functions in GraphAndDates
that create SVG elements. So now, they are separated which is more
practical for testing, refactoring, readings and so on. Also, I've moved
them to a separate file, so now GraphAndDates code looks more clear.
I've covered with tests all these functions, except from the main and
the most complex - createGraphElems, I started writing tests for it, but
I haven't finished yet.