Make some changes
Changed function declarations of some components, also fixed some bugs
and improved (I hope) rendering logic of SideNav, so it doesn't create
errors while SSR anymore, since it doesn't access window and document
objects. But there are problems with it's appearence after SSR, it
blinks, which is expected on CSR, but when there are also generated
content, it feels jaring. Since I can't check width of device on
server, it's impossible to define what kind of menu server should
deliver. Maybe, duration of blinking is noticeably because some
layoutEffects of other components take a lot of time. Well, I'm pretty
satisfied with the way SideNav works, I just need to figure out how to
make it more SSR or reduce time of blinking